requirements/local.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.0.6
django-bootstrap4==24.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-model-utils==5.0.0
pillow==10.4.0
psycopg2==2.9.9
setuptools==72.2.0
soupsieve==2.6
sqlparse==0.5.0
tzdata==2024.1
Unipath==1.1